+++ test/patch.js
@@ -350,6 +350,20 @@ describe('patch', function(){
       .expect('content-type', 'application/x-bogus; charset=utf-8', done);
     })
 
+    it('should not double-include charset in content-type', function(done){
+      var app = connect();
+
+      app.use(function(req, res){
+        res.charset = 'utf-8';
+        res.setHeader('content-type', 'application/x-bogus; charset=utf-8');
+        res.end('hello!');
+      })
+
+      app.request()
+      .get('/')
+      .expect('content-type', 'application/x-bogus; charset=utf-8', done);
+    })
+
     describe('should handle set-cookie setting multiple cookies as array', function(){
       it('should be emitted', function(done){
         var app = connect();